chore: extract hasRowChanged
