+++ cspell.json
@@ -12,6 +12,7 @@
         "Multicall",
         "nomodule",
         "nyquist",
+        "recyclerlistview",
         "stylelint",
         "Swipeable",
         "untrusted",
